requirements.txt
docxtpl
pydantic
num2words
google-auth-oauthlib
google-api-python-client
supabase
python-dotenv
email-validator
pytz
